# go-api/internal/infra/webserver/handles/user_handler.go
package handles

import (
	"encoding/json"
	"net/http"
	"time"

	"github.com/go-chi/jwtauth"
	"github.com/renan5g/go-expert/internal/dto"
	"github.com/renan5g/go-expert/internal/entity"
	"github.com/renan5g/go-expert/internal/infra/database"
)

type Error struct {
	Message string `json:"message"`
}

type UserHandle struct {
	UserDB database.UserInterface
}

func NewUserHandle(db database.UserInterface) *UserHandle {
	return &UserHandle{UserDB: db}
}

// GetJWT godoc
// @Summary 		Get a user JWT
// @Description Get a user JWT
// @Tags 				users
// @Accept 			json
// @Produce 		json
// @Param 			request body dto.GetJWTInput true "user credentials"
// @Success 		200 {object} dto.GetJWTOutput
// @Failure			404
// @Failure 		500 {object} Error
// @Router 			/users/generate_token [post]
func (h *UserHandle) GetJwt(w http.ResponseWriter, r *http.Request) {
	jwt := r.Context().Value("jwt").(*jwtauth.JWTAuth)
	jwtExpiriesIn := r.Context().Value("jwtExpiresIn").(int)

	var user dto.GetJWTInput
	err := json.NewDecoder(r.Body).Decode(&user)
	if err != nil {
		w.WriteHeader(http.StatusBadRequest)
		err := Error{Message: err.Error()}
		json.NewEncoder(w).Encode(err)
		return
	}

	u, err := h.UserDB.FindByEmail(user.Email)
	if err != nil {
		w.WriteHeader(http.StatusNotFound)
		err := Error{Message: err.Error()}
		json.NewEncoder(w).Encode(err)
		return
	}

	if !u.ValidatePassword(user.Password) {
		w.WriteHeader(http.StatusUnauthorized)
		err := Error{Message: err.Error()}
		json.NewEncoder(w).Encode(err)
		return
	}

	_, token, _ := jwt.Encode(map[string]interface{}{
		"sub": u.ID.String(),
		"exp": time.Now().Add(time.Second * time.Duration(jwtExpiriesIn)).Unix(),
	})
	accessToken := dto.GetJWTOutput{AccessToken: token}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(accessToken)
}

// Create user godoc
// @Summary 		Create user
// @Description Create user
// @Tags 				users
// @Accept 			json
// @Produce 		json
// @Param 			request body dto.CreateUserInput true "user request"
// @Success 		201
// @Failure 		500 {object} Error
// @Router 			/users [post]
func (h *UserHandle) Create(w http.ResponseWriter, r *http.Request) {
	var user dto.CreateUserInput
	err := json.NewDecoder(r.Body).Decode(&user)
	if err != nil {
		w.WriteHeader(http.StatusBadRequest)
		err := Error{Message: err.Error()}
		json.NewEncoder(w).Encode(err)
		return
	}

	u, err := entity.NewUser(user.Name, user.Email, user.Password)
	if err != nil {
		w.WriteHeader(http.StatusBadRequest)
		err := Error{Message: err.Error()}
		json.NewEncoder(w).Encode(err)
		return
	}

	err = h.UserDB.Create(u)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		err := Error{Message: err.Error()}
		json.NewEncoder(w).Encode(err)
		return
	}

	w.WriteHeader(http.StatusCreated)
}